Add support for incremental reading
This is intended to be used in a select()-like loop that
manages several file descriptors and can't block on any
single one to read an entire message. The reader will
wait to see which fd is ready to read, and then incrementally
read from it. If the entire message isn't ready though, the
state will be stored, and the reader can loop again to see
which other file descriptor is ready for reading.